Split into two test files
