+++ tests/common/test_monkey_patching.py
@@ -11,16 +11,7 @@
  limitations under the License.
 """
 
-import torch
-from torch import nn
-from torch.nn import functional as F
-
 from nncf.common.utils.patcher import PATCHER
-from nncf.torch import create_compressed_model
-from nncf.torch.dynamic_graph.context import get_current_context
-from nncf.torch.dynamic_graph.context import patch_method_with_no_nncf_trace
-from tests.torch.helpers import create_conv
-from tests.torch.helpers import get_empty_config
 
 
 CORRECT_WRAPPER_STACK = "base"
@@ -95,45 +86,3 @@ def test_patcher():
     PATCHER.patch(TestOverrideClass.assert_wrapper_stack_method, wrapper1, override=True)
     CORRECT_WRAPPER_STACK = "base_wrapper1"
     test_obj.assert_wrapper_stack_method(wrapper_stack="base")
-
-
-class TestModel(nn.Module):
-    """
-    A test model with an operation resulting in an ambiguous graph.
-    Ambiguous operation output is put into the model output for testing convenience.
-    """
-
-    def __init__(self, correct_is_tracing):
-        super().__init__()
-        self.conv = create_conv(in_channels=1, out_channels=1, kernel_size=2)
-        self.correct_is_tracing = correct_is_tracing
-
-    def forward(self, x):
-        x = F.sigmoid(self.conv(x - 0.5))
-        output = self.ambiguous_op(x)
-        return x, output
-
-    def ambiguous_op(self, x):
-        assert get_current_context().is_tracing == self.correct_is_tracing
-
-        output = torch.zeros_like(x)
-        for idx in range(torch.greater(x, 0.5).sum()):
-            output = output + x
-        return output
-
-
-def test_no_trace_model_patching():
-    config = get_empty_config()
-    config["input_info"] = {
-        "sample_size": [1, 1, 4, 4],
-        "filler": "random"
-    }
-
-    # Not patching anything: all output nodes are traced
-    _, compressed_model = create_compressed_model(TestModel(True), config)
-    assert len(compressed_model._original_graph._output_nncf_nodes) == 2
-
-    # Patching a function results with no_nncf_trace in method not producing an output node
-    patch_method_with_no_nncf_trace(TestModel.ambiguous_op)
-    _, compressed_model = create_compressed_model(TestModel(False), get_empty_config())
-    assert len(compressed_model._original_graph._output_nncf_nodes) == 1
